+++ client/src/App.tsx
@@ -9,7 +9,7 @@ import { SocketProvider } from './context/useSocketContext';
 import { SnackBarProvider } from './context/useSnackbarContext';
 
 import './App.css';
-import Bookings from './pages/ManageBookings/Bookings';
+import Bookings from './pages/Bookings/Bookings';
 
 function App(): JSX.Element {
   return (
@@ -23,7 +23,6 @@ function App(): JSX.Element {
                 <Route exact path="/signup" component={Signup} />
                 <Route exact path="/dashboard">
                   <Dashboard />
-                  {/* <Bookings /> */}
                 </Route>
                 <Route exact path="/mysitter" component={Bookings} />
                 <Route path="*">
