+++ client/src/pages/Bookings/Calendar.tsx
@@ -1,8 +1,4 @@
 import 'date-fns';
-import React from 'react';
-import Grid from '@material-ui/core/Grid';
-import DateFnsUtils from '@date-io/date-fns';
-import { MuiPickersUtilsProvider, KeyboardDatePicker } from '@material-ui/pickers';
 import useStyles from './useStyles';
 import MultipleDatesPicker from '@randex/material-ui-multiple-dates-picker';
 import Paper from '@material-ui/core/Paper';
